feat(query): add query params validation
Add validation class for query params. Perform validation for queries based on validation schema.
If a value match validation schema it's available in the match query props. If it's not valid the value is undefined.
If the validation schema requires some value, it will be populated if it's not matched the schema or it's not exists.